src/backend/requirements.txt
fastapi==0.105.*
Pillow==10.1.*
python-dotenv==1.0.*
redis==5.0.*
uvicorn==0.24.*